requirements.txt
pkg-resources==0.0.0
pynmea2==1.15.0
pyserial==3.4
